Chunk 9: split public.js into routes/public/ sub-routers
Refactor the 1,455-line routes/public.js into a sub-router directory
following the groups/members/portal precedent. Pure move, no behaviour
change:

- join.js       — online joining + PayPal flow (join-config, join,
                  payment-confirm, resume-payment, email-payment-link) plus
                  the join confirmation / officer notification email stubs
- portalAuth.js — unauthenticated portal credential endpoints (register,
                  verify-email, login, forgot-password, reset-password) with
                  the portal-auth rate limiter, lockout helper, and reset /
                  verification email senders
- read.js       — public groups and calendar information pages
- index.js      — owns the resolveTenant middleware, mounts the three
                  sub-routers, then mounts the authenticated portal app
                  router at /:slug/portal/app

app.js now imports public/index.js. Backend suite green (593 tests),
lint and Prettier clean.
