+++ CHANGELOG.md
@@ -113,10 +113,21 @@ Format: `## [version] — YYYY-MM-DD` with bullet points per change.
     list router before the `/:id` CRUD router so route matching order is
     preserved). No shared `helpers.js` was needed — each field-def/schema/access
     helper is local to one sub-router. `app.js` now imports `groups/index.js`.
+  - `routes/public.js` (1,455 lines) → `routes/public/` with `join.js` (the
+    online-joining and PayPal flow: join-config, join, payment-confirm,
+    resume-payment, email-payment-link, plus the join confirmation/officer
+    notification email stubs), `portalAuth.js` (the unauthenticated portal
+    credential endpoints register/verify-email/login/forgot-password/
+    reset-password, with the portal-auth rate limiter, lockout helper, and
+    reset/verification email senders), `read.js` (the public groups and
+    calendar information pages), and `index.js` (owns the `resolveTenant`
+    middleware, mounts the three sub-routers, then mounts the authenticated
+    portal app router at `/:slug/portal/app`). `app.js` now imports
+    `public/index.js`.
   - Route registrations verified identical before/after; backend suite green
     (45 files, 593 tests), lint and Prettier clean. The remaining oversized
-    routes (`public.js`, `teams.js`, `calendar.js`) are deferred to
-    follow-up sessions per the chunk's "one file per session" note.
+    routes (`teams.js`, `calendar.js`) are deferred to follow-up sessions per
+    the chunk's "one file per session" note.
 
 ### Fixed
 - **Duplicated security findings in `KNOWN-ISSUES.md`** — the Chunk 4 and Chunk 5
