+++ backend/src/app.js
@@ -34,7 +34,7 @@ import emailRoutes from './routes/email.js';
 import giftAidRoutes from './routes/giftAid.js';
 import systemMessageRoutes from './routes/systemMessages.js';
 import publicLinksRoutes from './routes/publicLinks.js';
-import publicRoutes from './routes/public.js';
+import publicRoutes from './routes/public/index.js';
 import calendarRoutes from './routes/calendar.js';
 import eventTypeRoutes from './routes/eventTypes.js';
 import membershipCardRoutes from './routes/membershipCards.js';
